refactor: remove references to LSP
