+++ src/hoverifier.test.ts
@@ -17,7 +17,7 @@ import {
 import { HoverOverlayProps } from './HoverOverlay'
 import { findPositionsFromEvents, SupportedMouseEvent } from './positions'
 import { CodeViewProps, DOM } from './testutils/dom'
-import { createHoverAttachment, createStubActionsFetcher, createStubHoverFetcher } from './testutils/lsp'
+import { createHoverAttachment, createStubActionsFetcher, createStubHoverFetcher } from './testutils/fixtures'
 import { dispatchMouseEventAtPositionImpure } from './testutils/mouse'
 import { LOADING } from './types'
 
@@ -578,9 +578,10 @@ describe('Hoverifier', () => {
                 )
 
                 const inputDiagram = 'ab'
-                // There is probably a bug in code that is unrelated to this feature that is causing the PositionAdjuster to be called an extra time.
-                // It should look like '-(ba)'. That is, we adjust the position from CodeViewToActual for the LSP fetches and then back from CodeViewToActual
-                // for highlighting the token in the DOM.
+                // There is probably a bug in code that is unrelated to this feature that is causing the
+                // PositionAdjuster to be called an extra time. It should look like '-(ba)'. That is, we adjust the
+                // position from CodeViewToActual for the fetches and then back from CodeViewToActual for
+                // highlighting the token in the DOM.
                 const outputDiagram = 'a(ba)'
 
                 const outputValues: {
